requirements.txt
opencv-python-headless
camelot-py
ghostscript
streamlit
pytesseract
tesseract-ocr
